Move components to server dir
