--- .gitignore
@@ -1,3 +1,0 @@
-.env
-.idea
-venv
